place the sheets creation in sbksheets.py
