+++ charts/sbk-charts.py
@@ -14,51 +14,9 @@ import argparse
 from collections import OrderedDict
 
 import openpyxl
-import pandas
-import xlsxwriter
 from openpyxl.chart import LineChart, Reference, Series
 from openpyxl.utils import get_column_letter
-
-
-class SbkSheets:
-    def __init__(self, iFile, oFile):
-        self.iFile = iFile
-        self.oFile = oFile
-
-    def wb_add_two_sheets(self, wb, r_name, t_name, df):
-        header = df.columns.values
-        r_ws = wb.add_worksheet(r_name)
-        t_ws = wb.add_worksheet(t_name)
-        for c, h in enumerate(header):
-            r_ws.set_column(c, c, len(h))
-            t_ws.set_column(c, c, len(h))
-            r_ws.write(0, c, h)
-            t_ws.write(0, c, h)
-        r_row = 1
-        t_row = 1
-        for row in df.iterrows():
-            if row[1]['Type'] == 'Total':
-                for c, h in enumerate(header):
-                    col_size = len(str(row[1][h])) + 1
-                    if col_size > len(h):
-                        t_ws.set_column(c, c, col_size)
-                    t_ws.write(t_row, c, row[1][h])
-                t_row += 1
-            else:
-                for c, h in enumerate(header):
-                    col_size = len(str(row[1][h])) + 1
-                    if col_size > len(h):
-                        r_ws.set_column(c, c, col_size)
-                    r_ws.write(r_row, c, row[1][h])
-                r_row += 1
-
-    def create_sheets(self):
-        df = pandas.read_csv(self.iFile)
-        wb = xlsxwriter.Workbook(self.oFile)
-        self.wb_add_two_sheets(wb, "R1", "T1", df)
-        wb.close()
-        print("xlsx file %s created" % self.oFile)
-
+import sbksheets
 
 class SbkCharts:
     def __init__(self, file):
@@ -212,7 +170,7 @@ def main():
     args = parser.parse_args()
     print('Input file is ', args.ifile)
     print('Output file is ', args.ofile)
-    sheets = SbkSheets(args.ifile, args.ofile)
+    sheets = sbksheets.SbkSheets(args.ifile, args.ofile)
     sheets.create_sheets()
     charts = SbkCharts(args.ofile)
     charts.create_graphs()
